+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,5 +1,6 @@
 ﻿### Version 3.8.12
 
+- Improvement: Better handling of `nullable` child entities when generating *create* and *update* handlers
 - Improvement: `WithoutMethodModifier` added to CSharpClassMethod to remvoe any method modifiers (e.g. `new`, `override`, `static`, `abstract`)
 - Improvement: `Lhs` and `Rhs` values of `CSharpObjectInitStatement` can now be read and modified.
 - Fixed: `NormalizeNamespace` has been fixed to ensure that it parses C# types correctly.